Fix create basic info label. Add digits when create basic. Fix amount, contract returns in fraction, thus need /100 to be actuall value; all user input should *100 before passing to contract.
