+++ src/app/pipes/time-duration.pipe.ts
@@ -1,0 +1,25 @@
+import { Pipe, PipeTransform } from '@angular/core';
+
+@Pipe({
+  name: 'timeDuration'
+})
+export class TimeDurationPipe implements PipeTransform {
+
+  transform(ms: number): string {
+    if (ms < 0) {
+      ms = -ms;
+    }
+    const time = {
+      day: Math.floor(ms / 86400000),
+      hour: Math.floor(ms / 3600000) % 24,
+      minute: Math.floor(ms / 60000) % 60,
+      second: Math.floor(ms / 1000) % 60,
+      millisecond: Math.floor(ms) % 1000
+    };
+    return Object.entries(time)
+      .filter(val => val[1] !== 0)
+      .map(([key, val]) => `${val} ${key}${val !== 1 ? 's' : ''}`)
+      .join(', ');
+  }
+
+}
